mejoro manejo de creacion de clientes
