+++ frontend/src/components/ClientsContainer/ClientsCreateModal.tsx
@@ -1,22 +1,38 @@
 // ClientsCreateModal.tsx
 import React from 'react';
 import { Modal } from 'react-bootstrap';
-import { ClientsCreate } from './ClientsCreate';
-import { ClienteCreate, ConfiguracionCreate, DetallesAdicionalesForUI } from '../../codegen_output';
+import { ClientesService, ClienteCreate, ConfiguracionCreate, DetallesAdicionalesForUI } from '../../codegen_output';
+import ClientsCreate from './ClientsCreate';
+import Swal from 'sweetalert2';
 
 interface ClientsCreateModalProps {
   show: boolean;
   onHide: () => void;
-  onNewClient: (
+  onClientAdded: () => void;  // This prop triggers a refresh in the parent.
+  expanded?: boolean;
+}
+
+const ClientsCreateModal: React.FC<ClientsCreateModalProps> = ({ show, onHide, onClientAdded, expanded = false }) => {
+  const handleNewClient = async (
     newClient: ClienteCreate, 
     tarjetaId: number, 
     additionalDetails?: DetallesAdicionalesForUI,
     maxAmounts?: ConfiguracionCreate
-    ) => Promise<void>;
-  expanded?: boolean;
-}
+  ) => {
+    try {
+      await ClientesService.handleCreateClienteWithTarjetaBackendApiV1ClientesPost(tarjetaId, {
+        cliente_in: newClient,
+        detalle_adicional_in: additionalDetails,
+        montos_config_in: maxAmounts,
+      });
+      Swal.fire(`${newClient.nombre}`, 'ha sido guardado con éxito', 'success');
+      onHide();  // Close modal after successful client creation
+      onClientAdded();  // Notify parent component to refresh data
+    } catch (error) {
+      Swal.fire('Error', 'No se pudo guardar el cliente.', 'error');
+    }
+  };
 
-const ClientsCreateModal: React.FC<ClientsCreateModalProps> = ({ show, onHide, onNewClient, expanded = false }) => {
   return (
     <Modal
       show={show}
@@ -32,11 +48,9 @@ const ClientsCreateModal: React.FC<ClientsCreateModalProps> = ({ show, onHide, o
         </Modal.Title>
       </Modal.Header>
       <Modal.Body className="new-client-modal-body">
-        <ClientsCreate onNewClient={onNewClient} expanded={expanded} />
+        <ClientsCreate onNewClient={handleNewClient} expanded={expanded} />
       </Modal.Body>
       <Modal.Footer className="new-client-modal-footer">
-        {/* <Button variant="secondary" onClick={handleClose}>Cerrar</Button> */}
-        {/* <Button className="new-client-modal-footer-btn" onClick={handleClose}>Dar de alta</Button> */}
       </Modal.Footer>
     </Modal>
   );
